requirements.txt
Flask==1.1.1
Flask-Admin==1.5.5
Flask-SQLAlchemy==2.4.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
Pillow